Create generic lora device driver based on an extended radio trait
