+++ device/Cargo.toml
@@ -59,7 +59,7 @@ env_logger = "0.8"
 default = [ "std", "log", "time" ]
 std = ["embassy/std"]
 time = ["embassy/time"]
-"lora+sx127x" = ["lorawan-device", "lorawan-encoding", "bit_field"]
+"lora+sx127x" = ["lorawan-device", "lorawan-encoding", "bit_field", "lora"]
 "lora+rak811" = ["nom", "moveslice"]
 "wifi+esp8266" = ["nom", "moveslice"]
 lora = []
