+++ device/src/drivers/lora/mod.rs
@@ -2,3 +2,9 @@
 pub mod rak811;
 #[cfg(feature = "lora+sx127x")]
 pub mod sx127x;
+
+#[cfg(feature = "lora")]
+pub mod device;
+
+#[cfg(feature = "lora")]
+pub use device::*;
