+++ device/src/drivers/lora/sx127x/mod.rs
@@ -1,414 +1,258 @@
-use crate::traits::lora::*;
-use core::future::Future;
-use embassy::time::*;
-use embassy::traits::gpio::WaitForRisingEdge;
-use embedded_hal::{
-    blocking::spi::{Transfer, Write},
-    digital::v2::OutputPin,
-};
-
+use super::{Radio, RadioPhyEvent};
+use crate::traits::lora::LoraError as DriverError;
+use embedded_hal::blocking::spi::{Transfer, Write};
+use embedded_hal::digital::v2::OutputPin;
 use lorawan_device::{
-    radio, region, Device as LorawanDevice, Error as LorawanError, Event as LorawanEvent,
-    Response as LorawanResponse,
+    radio::{
+        Bandwidth, Error as LoraError, Event as LoraEvent, PhyRxTx, Response as LoraResponse,
+        RxQuality, SpreadingFactor,
+    },
+    Timings,
 };
-use lorawan_encoding::default_crypto::DefaultFactory as Crypto;
 
 mod sx127x_lora;
-mod sx127x_radio;
+use sx127x_lora::{LoRa, RadioMode, IRQ};
 
-use sx127x_radio::{RadioPhyEvent, Sx127xRadio as Radio};
-
-enum DriverState<'a, SPI, CS, RESET, E>
+pub struct Sx127xRadio<'a, SPI, CS, RESET, E>
 where
-    SPI: Transfer<u8, Error = E> + Write<u8, Error = E> + 'static,
-    CS: OutputPin + 'static,
-    RESET: OutputPin + 'static,
-    E: 'static,
+    SPI: Transfer<u8, Error = E> + Write<u8, Error = E>,
+    CS: OutputPin,
+    RESET: OutputPin,
 {
-    New(Radio<'a, SPI, CS, RESET, E>, &'a mut [u8]),
-    Configured(LorawanDevice<'a, Radio<'a, SPI, CS, RESET, E>, Crypto>),
+    radio: LoRa<SPI, CS, RESET>,
+    radio_state: State,
+    rx_buffer: &'a mut [u8],
+    rx_buffer_written: usize,
 }
 
-pub struct Sx127xDriver<'a, P, SPI, CS, RESET, E>
-where
-    P: WaitForRisingEdge,
-    SPI: Transfer<u8, Error = E> + Write<u8, Error = E> + 'static,
-    CS: OutputPin + 'static,
-    RESET: OutputPin + 'static,
-    E: 'static,
-{
-    irq: P,
-    state: Option<DriverState<'a, SPI, CS, RESET, E>>,
-    get_random: fn() -> u32,
+#[derive(Debug, Copy, Clone)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum State {
+    Idle,
+    Txing,
+    Rxing,
 }
 
-pub enum DriverEvent {
-    ProcessAfter(u32),
-    JoinSuccess,
-    JoinFailed,
-    SessionExpired,
-    Ack,
-    AckWithData(usize, [u8; 255]),
-    AckTimeout,
-    None,
+fn spreading_factor_to_u8(sf: SpreadingFactor) -> u8 {
+    match sf {
+        SpreadingFactor::_7 => 7,
+        SpreadingFactor::_8 => 8,
+        SpreadingFactor::_9 => 9,
+        SpreadingFactor::_10 => 10,
+        SpreadingFactor::_11 => 11,
+        SpreadingFactor::_12 => 12,
+    }
 }
 
-impl<'a, P, SPI, CS, RESET, E> Sx127xDriver<'a, P, SPI, CS, RESET, E>
+fn bandwidth_to_i64(bw: Bandwidth) -> i64 {
+    match bw {
+        Bandwidth::_125KHz => 125_000,
+        Bandwidth::_250KHz => 250_000,
+        Bandwidth::_500KHz => 500_000,
+    }
+}
+
+impl<'a, SPI, CS, RESET, E> Sx127xRadio<'a, SPI, CS, RESET, E>
 where
-    P: WaitForRisingEdge,
-    SPI: Transfer<u8, Error = E> + Write<u8, Error = E> + 'a,
-    CS: OutputPin + 'a,
-    RESET: OutputPin + 'a,
+    SPI: Transfer<u8, Error = E> + Write<u8, Error = E>,
+    CS: OutputPin,
+    RESET: OutputPin,
 {
-    pub fn new(
-        irq: P,
-        spi: SPI,
-        cs: CS,
-        reset: RESET,
-        get_random: fn() -> u32,
-        radio_tx_buf: &'a mut [u8],
-        radio_rx_buf: &'a mut [u8],
-    ) -> Self {
-        let radio = Radio::new(spi, cs, reset, radio_rx_buf);
+    pub fn new(spi: SPI, cs: CS, reset: RESET, rx_buffer: &'a mut [u8]) -> Self {
         Self {
-            irq,
-            state: Some(DriverState::New(radio, radio_tx_buf)),
-            get_random,
+            radio_state: State::Idle,
+            radio: LoRa::new(spi, cs, reset),
+            rx_buffer,
+            rx_buffer_written: 0,
         }
     }
 
-    fn process_event(
+    pub fn handle_event_idle(
         &mut self,
-        event: LorawanEvent<'a, Radio<'a, SPI, CS, RESET, E>>,
-    ) -> DriverEvent {
-        //crate::log_stack("Process event");
-        match self.state.take().unwrap() {
-            DriverState::Configured(lorawan) => {
-                match &event {
-                    LorawanEvent::NewSessionRequest => {
-                        trace!("New Session Request");
-                    }
-                    LorawanEvent::RadioEvent(e) => match e {
-                        radio::Event::TxRequest(_, _) => (),
-                        radio::Event::RxRequest(_) => (),
-                        radio::Event::CancelRx => (),
-                        radio::Event::PhyEvent(_) => {
-                            trace!("Phy event");
-                        }
-                    },
-                    LorawanEvent::TimeoutFired => (),
-                    LorawanEvent::SendDataRequest(_e) => {
-                        trace!("SendData");
-                    }
-                }
-                let (mut new_state, response) = lorawan.handle_event(event);
-                trace!("Event handled");
-                let event = self.process_response(&mut new_state, response);
-                self.state.replace(DriverState::Configured(new_state));
-                event
-            }
-            s => {
-                trace!("Not yet configured, event processing skipped");
-                self.state.replace(s);
-                DriverEvent::None
-            }
-        }
-    }
+        event: LoraEvent<Self>,
+    ) -> (State, Result<LoraResponse<Self>, LoraError<Self>>) {
+        match event {
+            LoraEvent::TxRequest(config, buf) => {
+                //trace!("Set config: {:?}", config);
+                let result = (move || {
+                    self.radio.set_tx_power(14, 0)?;
+                    self.radio.set_frequency(config.rf.frequency)?;
+                    // TODO: Modify radio to support other coding rates
+                    self.radio.set_coding_rate_4(5)?;
+                    self.radio
+                        .set_signal_bandwidth(bandwidth_to_i64(config.rf.bandwidth))?;
+                    self.radio
+                        .set_spreading_factor(spreading_factor_to_u8(config.rf.spreading_factor))?;
 
-    fn process_response<'m>(
-        &self,
-        lorawan: &mut LorawanDevice<'m, Radio<'m, SPI, CS, RESET, E>, Crypto>,
-        response: Result<LorawanResponse, LorawanError<Radio<'m, SPI, CS, RESET, E>>>,
-    ) -> DriverEvent {
-        //crate::log_stack("Process response");
-        match response {
-            Ok(response) => match response {
-                LorawanResponse::TimeoutRequest(ms) => {
-                    trace!("TimeoutRequest: {:?}", ms);
-                    return DriverEvent::ProcessAfter(ms);
-                }
-                LorawanResponse::JoinSuccess => {
-                    return DriverEvent::JoinSuccess;
-                }
-                LorawanResponse::ReadyToSend => {
-                    trace!("RxWindow expired but no ACK expected. Ready to Send");
-                }
-                LorawanResponse::DownlinkReceived(fcnt_down) => {
-                    if let Some(downlink) = lorawan.take_data_downlink() {
-                        use lorawan_encoding::parser::FRMPayload;
+                    self.radio.set_preamble_length(8)?;
+                    self.radio.set_lora_pa_ramp()?;
+                    self.radio.set_lora_sync_word()?;
+                    self.radio.set_invert_iq(false)?;
+                    self.radio.set_crc(true)?;
 
-                        if let Ok(FRMPayload::Data(data)) = downlink.frm_payload() {
-                            trace!(
-                                "Downlink received \t\t(FCntDown={}\tFRM: {:?})",
-                                fcnt_down,
-                                data,
-                            );
-                            let mut buf = [0; 255];
-                            buf[0..data.len()].copy_from_slice(&data[0..data.len()]);
-                            return DriverEvent::AckWithData(data.len(), buf);
-                        } else {
-                            trace!("Downlink received \t\t(FcntDown={})", fcnt_down);
-                            return DriverEvent::Ack;
-                        }
-
-                        /*
-                        let fhdr = downlink.fhdr();
-                        let fopts = fhdr.fopts();
-                        let mut mac_commands_len = 0;
-                        for mac_command in fopts {
-                            if mac_commands_len == 0 {
-                                trace!("\tFOpts: ");
-                            }
-                            trace!("{:?},", mac_command);
-                            mac_commands_len += 1;
-                        }
-                        */
-                    }
-                }
-                LorawanResponse::NoAck => {
-                    trace!("RxWindow expired, expected ACK to confirmed uplink not received");
-                    return DriverEvent::AckTimeout;
-                }
-                LorawanResponse::NoJoinAccept => {
-                    trace!("No Join Accept Received. Retrying.");
-                    return DriverEvent::JoinFailed;
-                }
-                LorawanResponse::SessionExpired => {
-                    trace!("SessionExpired. Created new Session");
-                    return DriverEvent::SessionExpired;
-                }
-                LorawanResponse::NoUpdate => {
-                    // info!("No update");
-                }
-                LorawanResponse::UplinkSending(fcnt_up) => {
-                    trace!("Uplink with FCnt {}", fcnt_up);
-                }
-                LorawanResponse::JoinRequestSending => {
-                    trace!("Join Request Sending");
-                }
-            },
-            Err(err) => match err {
-                LorawanError::Radio(_) => error!("Radio error"),
-                LorawanError::Session(_) => error!("Session error"), //{:?}", e),
-                LorawanError::NoSession(_) => error!("NoSession error"),
-            },
-        }
-        DriverEvent::None
-    }
-
-    async fn join(&mut self) -> Result<(), LoraError> {
-        //crate::log_stack("Driver join");
-        let mut event: DriverEvent = self.process_event(LorawanEvent::NewSessionRequest);
-        loop {
-            match event {
-                DriverEvent::ProcessAfter(ms) => {
-                    let interrupt = self.irq.wait_for_rising_edge();
-                    match with_timeout(Duration::from_millis(ms.into()), interrupt).await {
-                        Ok(_) => {
-                            event = self.process_event(LorawanEvent::RadioEvent(
-                                radio::Event::PhyEvent(RadioPhyEvent::Irq),
-                            ));
-                        }
-                        Err(TimeoutError) => {
-                            event = self.process_event(LorawanEvent::TimeoutFired);
-                        }
-                    }
-                }
-                DriverEvent::JoinSuccess => {
-                    trace!("Joined successfully");
-                    return Ok(());
-                }
-                DriverEvent::JoinFailed => {
-                    event = self.process_event(LorawanEvent::NewSessionRequest);
-                }
-                _ => {
-                    // Wait for interrupt
-                    self.irq.wait_for_rising_edge().await;
-                    event = self.process_event(LorawanEvent::RadioEvent(radio::Event::PhyEvent(
-                        RadioPhyEvent::Irq,
-                    )));
+                    self.radio.set_dio0_tx_done()?;
+                    self.radio.transmit_payload(&buf[..])
+                })();
+                match result {
+                    Ok(_) => (State::Txing, Ok(LoraResponse::Txing)),
+                    Err(_) => (State::Idle, Err(LoraError::PhyError(()))),
                 }
             }
-        }
-    }
+            LoraEvent::RxRequest(config) => {
+                // trace!("Set RX config: {:?}", config);
+                let result = (move || {
+                    self.radio.reset_payload_length()?;
+                    self.radio.set_frequency(config.frequency)?;
+                    // TODO: Modify radio to support other coding rates
+                    self.radio.set_coding_rate_4(5)?;
+                    self.radio
+                        .set_signal_bandwidth(bandwidth_to_i64(config.bandwidth))?;
+                    self.radio
+                        .set_spreading_factor(spreading_factor_to_u8(config.spreading_factor))?;
 
-    async fn send_data(
-        &mut self,
-        qos: QoS,
-        port: Port,
-        data: &[u8],
-    ) -> Result<DriverEvent, LoraError> {
-        match self.state.take().unwrap() {
-            DriverState::Configured(lorawan) => {
-                let ready_to_send = lorawan.ready_to_send_data();
-                if ready_to_send {
-                    let (mut new_state, response) = lorawan.send(
-                        data,
-                        port,
-                        match qos {
-                            QoS::Confirmed => true,
-                            QoS::Unconfirmed => false,
-                        },
+                    self.radio.set_preamble_length(8)?;
+                    self.radio.set_lora_sync_word()?;
+                    self.radio.set_invert_iq(true)?;
+                    self.radio.set_crc(true)?;
+
+                    self.radio.set_dio0_rx_done()?;
+                    self.radio.set_mode(RadioMode::RxContinuous)
+
+                    /*
+                    let irq_flags = self.radio.irq_flags().ok().unwrap();
+                    let irq_flags_mask = self.radio.irq_flags_mask().ok().unwrap();
+                    info!(
+                        "RX STARTED, IRQ Flags: 0x{:x}, Mask: 0x{:x}",
+                        irq_flags,
+                        irq_flags_mask
                     );
-                    let event = self.process_response(&mut new_state, response);
-                    self.state.replace(DriverState::Configured(new_state));
-                    Ok(event)
-                } else {
-                    self.state.replace(DriverState::Configured(lorawan));
-                    Err(LoraError::NotReady)
+                    r*/
+                })();
+                match result {
+                    Ok(_) => (State::Rxing, Ok(LoraResponse::Rxing)),
+                    Err(_) => (State::Rxing, Err(LoraError::PhyError(()))),
                 }
             }
-            other => {
-                //info!("Driver not yet initialized, ignoring configuration");
-                self.state.replace(other);
-                Err(LoraError::OtherError)
-            }
+            // deny any events while idle; they are unexpected
+            LoraEvent::PhyEvent(_) => (State::Idle, Err(LoraError::PhyError(()))),
+            LoraEvent::CancelRx => (State::Idle, Err(LoraError::CancelRxWhileIdle)),
         }
     }
 
-    async fn send_recv(
+    pub fn handle_event_txing(
         &mut self,
-        qos: QoS,
-        port: Port,
-        data: &[u8],
-        rx: Option<&mut [u8]>,
-    ) -> Result<usize, LoraError> {
-        // Await response
-        let mut event = self.send_data(qos, port, data).await?;
-        loop {
-            match event {
-                DriverEvent::ProcessAfter(ms) => {
-                    let interrupt = self.irq.wait_for_rising_edge();
-                    match with_timeout(Duration::from_millis(ms.into()), interrupt).await {
-                        Ok(_) => {
-                            event = self.process_event(LorawanEvent::RadioEvent(
-                                radio::Event::PhyEvent(RadioPhyEvent::Irq),
-                            ));
-                        }
-                        Err(TimeoutError) => {
-                            event = self.process_event(LorawanEvent::TimeoutFired);
-                        }
+        event: LoraEvent<Self>,
+    ) -> (State, Result<LoraResponse<Self>, LoraError<Self>>) {
+        match event {
+            LoraEvent::PhyEvent(phyevent) => match phyevent {
+                RadioPhyEvent::Irq => {
+                    self.radio.set_mode(RadioMode::Stdby).ok().unwrap();
+                    let irq = self.radio.clear_irq().ok().unwrap();
+                    if (irq & IRQ::IrqTxDoneMask.addr()) != 0 {
+                        (State::Idle, Ok(LoraResponse::TxDone(0)))
+                    } else {
+                        (State::Txing, Ok(LoraResponse::Txing))
                     }
                 }
-                DriverEvent::AckWithData(len, buf) => {
-                    trace!("Received {} bytes of data", len);
-                    if let Some(rx) = rx {
-                        rx[0..len].copy_from_slice(&buf[0..len]);
+            },
+            LoraEvent::TxRequest(_, _) => (State::Txing, Err(LoraError::TxRequestDuringTx)),
+            LoraEvent::RxRequest(_) => (State::Txing, Err(LoraError::RxRequestDuringTx)),
+            LoraEvent::CancelRx => (State::Txing, Err(LoraError::CancelRxDuringTx)),
+        }
+    }
+
+    pub fn handle_event_rxing(
+        &mut self,
+        event: LoraEvent<Self>,
+    ) -> (State, Result<LoraResponse<Self>, LoraError<Self>>) {
+        match event {
+            LoraEvent::PhyEvent(phyevent) => match phyevent {
+                RadioPhyEvent::Irq => {
+                    self.radio.set_mode(RadioMode::Stdby).ok().unwrap();
+                    let irq = self.radio.clear_irq().ok().unwrap();
+                    if (irq & IRQ::IrqRxDoneMask.addr()) != 0 {
+                        let rssi = self.radio.get_packet_rssi().unwrap_or(0) as i16;
+                        let snr = self.radio.get_packet_snr().unwrap_or(0.0) as i8;
+                        if let Ok(size) = self.radio.read_packet_size() {
+                            self.radio.read_packet(self.rx_buffer).ok().unwrap();
+                            self.rx_buffer_written = size;
+                        }
+                        self.radio.set_mode(RadioMode::Sleep).ok().unwrap();
+                        (
+                            State::Idle,
+                            Ok(LoraResponse::RxDone(RxQuality::new(rssi, snr))),
+                        )
+                    } else {
+                        (State::Rxing, Ok(LoraResponse::Rxing))
                     }
-                    return Ok(len);
                 }
-                DriverEvent::AckTimeout => {
-                    trace!("Ack timed out!");
-                    return Err(LoraError::AckTimeout);
-                }
-                DriverEvent::Ack => {
-                    trace!("Ack received!");
-                    return Ok(0);
-                }
-                _ => {
-                    // Wait for interrupt
-                    self.irq.wait_for_rising_edge().await;
-                    event = self.process_event(LorawanEvent::RadioEvent(radio::Event::PhyEvent(
-                        RadioPhyEvent::Irq,
-                    )));
-                }
+            },
+            LoraEvent::CancelRx => {
+                self.radio.set_mode(RadioMode::Sleep).ok().unwrap();
+                (State::Idle, Ok(LoraResponse::Idle))
             }
+            LoraEvent::TxRequest(_, _) => (State::Rxing, Err(LoraError::TxRequestDuringTx)),
+            LoraEvent::RxRequest(_) => (State::Rxing, Err(LoraError::RxRequestDuringRx)),
         }
     }
 }
 
-impl<'a, P, SPI, CS, RESET, E> LoraDriver for Sx127xDriver<'a, P, SPI, CS, RESET, E>
+impl<'a, SPI, CS, RESET, E> Timings for Sx127xRadio<'a, SPI, CS, RESET, E>
 where
-    P: WaitForRisingEdge + 'a,
-    SPI: Transfer<u8, Error = E> + Write<u8, Error = E> + 'a,
-    E: 'a,
-    CS: OutputPin + 'a,
-    RESET: OutputPin + 'a,
+    SPI: Transfer<u8, Error = E> + Write<u8, Error = E>,
+    CS: OutputPin,
+    RESET: OutputPin,
 {
-    #[rustfmt::skip]
-    type ConfigureFuture<'m> where 'a: 'm = impl Future<Output = Result<(), LoraError>> + 'm;
-    fn configure<'m>(&'m mut self, config: &'m LoraConfig) -> Self::ConfigureFuture<'m> {
-        async move {
-            match self.state.take().unwrap() {
-                DriverState::New(mut radio, radio_tx_buf) => {
-                    //crate::log_stack("lora driver configure");
-                    radio.reset().await?;
-                    //info!("Configuring radio");
-                    let dev_eui = config.device_eui.as_ref().expect("device EUI must be set");
-                    let app_eui = config.app_eui.as_ref().expect("app EUI must be set");
-                    let app_key = config.app_key.as_ref().expect("app KEY must be set");
-                    //info!("Creating device");
-                    let data_rate =
-                        to_datarate(config.spreading_factor.unwrap_or(SpreadingFactor::SF9));
-                    let region = to_region(config.region.unwrap_or(LoraRegion::EU868));
-                    if let Err(e) = region {
-                        return Err(e);
-                    }
-                    let mut region = region.unwrap();
-                    region.set_receive_delay1(5000);
-                    let mut lorawan = LorawanDevice::new(
-                        region,
-                        radio,
-                        dev_eui.reverse().into(),
-                        app_eui.reverse().into(),
-                        app_key.clone().into(),
-                        self.get_random,
-                        radio_tx_buf,
-                    );
-                    lorawan.set_datarate(data_rate);
-                    self.state.replace(DriverState::Configured(lorawan));
-                    Ok(())
-                }
-                other => {
-                    //info!("Driver not yet initialized, ignoring configuration");
-                    self.state.replace(other);
-                    Err(LoraError::OtherError)
-                }
-            }
-        }
+    fn get_rx_window_offset_ms(&self) -> i32 {
+        -500
     }
-
-    #[rustfmt::skip]
-    type JoinFuture<'m> where 'a: 'm = impl Future<Output = Result<(), LoraError>> + 'm;
-    fn join<'m>(&'m mut self, _: ConnectMode) -> Self::JoinFuture<'m> {
-        async move { self.join().await }
-    }
-
-    #[rustfmt::skip]
-    type SendFuture<'m> where 'a: 'm = impl Future<Output = Result<(), LoraError>> + 'm;
-    fn send<'m>(&'m mut self, qos: QoS, port: Port, data: &'m [u8]) -> Self::SendFuture<'m> {
-        async move { self.send_recv(qos, port, data, None).await.map(|_| ()) }
-    }
-
-    #[rustfmt::skip]
-    type SendRecvFuture<'m> where 'a: 'm = impl Future<Output = Result<usize, LoraError>> + 'm;
-    fn send_recv<'m>(
-        &'m mut self,
-        qos: QoS,
-        port: Port,
-        data: &'m [u8],
-        rx: &'m mut [u8],
-    ) -> Self::SendRecvFuture<'m> {
-        async move { self.send_recv(qos, port, data, Some(rx)).await }
+    fn get_rx_window_duration_ms(&self) -> u32 {
+        800
     }
 }
 
-fn to_region(region: LoraRegion) -> Result<region::Configuration, LoraError> {
-    match region {
-        LoraRegion::EU868 => Ok(region::EU868::default().into()),
-        LoraRegion::US915 => Ok(region::US915::default().into()),
-        LoraRegion::CN470 => Ok(region::CN470::default().into()),
-        _ => Err(LoraError::UnsupportedRegion),
+impl<'a, SPI, CS, RESET, E> PhyRxTx for Sx127xRadio<'a, SPI, CS, RESET, E>
+where
+    SPI: Transfer<u8, Error = E> + Write<u8, Error = E>,
+    CS: OutputPin,
+    RESET: OutputPin,
+{
+    type PhyEvent = RadioPhyEvent;
+    type PhyError = ();
+    type PhyResponse = ();
+
+    fn get_mut_radio(&mut self) -> &mut Self {
+        self
+    }
+
+    fn get_received_packet(&mut self) -> &mut [u8] {
+        &mut self.rx_buffer[..self.rx_buffer_written]
+    }
+
+    fn handle_event(
+        &mut self,
+        event: LoraEvent<Self>,
+    ) -> Result<LoraResponse<Self>, LoraError<Self>>
+    where
+        Self: core::marker::Sized,
+    {
+        let (new_state, response) = match &self.radio_state {
+            State::Idle => self.handle_event_idle(event),
+            State::Txing => self.handle_event_txing(event),
+            State::Rxing => self.handle_event_rxing(event),
+        };
+        self.radio_state = new_state;
+        response
     }
 }
 
-fn to_datarate(spreading_factor: SpreadingFactor) -> region::DR {
-    match spreading_factor {
-        SpreadingFactor::SF7 => region::DR::_5,
-        SpreadingFactor::SF8 => region::DR::_4,
-        SpreadingFactor::SF9 => region::DR::_3,
-        SpreadingFactor::SF10 => region::DR::_2,
-        SpreadingFactor::SF11 => region::DR::_1,
-        SpreadingFactor::SF12 => region::DR::_0,
+impl<'a, SPI, CS, RESET, E> Radio for Sx127xRadio<'a, SPI, CS, RESET, E>
+where
+    SPI: Transfer<u8, Error = E> + Write<u8, Error = E>,
+    CS: OutputPin,
+    RESET: OutputPin,
+{
+    fn reset(&mut self) -> Result<(), DriverError> {
+        self.radio.reset().map_err(|_| DriverError::OtherError)
     }
 }
